Suggest some fine-tuning for foreign PR
In [PR_127] we got miscellaneous improvements. But we do no longer use
poll and therefore are back to the FD_SETSIZE limit of *select*.

This commit suggests some fine-tuning to [PR_127]. It does so by
replacing a possible SEGFAULT (aka crashing the whole JVM) by throwing a
java exception instead. This improves the reported error and even
enables user code to react to the situation.

Still not perfect. But IMHO good enough for now. We can still improve
later (for example as soon there's enough time to do so).

[PR_127]: https://github.com/java-native/jssc/pull/127