Make writeBytes() return how many bytes actually got written
The core change is this one:

--- src/main/java/jssc/SerialNativeInterface.java
+++ src/main/java/jssc/SerialNativeInterface.java
@@ -278,5 +278,5 @@ public class SerialNativeInterface {
      * @return If the operation is successfully completed, the method returns true, otherwise false
      */
-    public native boolean writeBytes(long handle, byte[] buffer) throws IOException;
+    public native int     writeBytes(long handle, byte[] buffer) throws IOException;

     /**

Essentially changing the return type of `writeBytes()` from `boolean` to
`int`. The reason behind this is that we can return how many bytes
actually got written, in place of just true/false.

Basically, this is a very small change. Nevertheless, there are some
consequences.

This is (kind of) a breaking change for the java API. But still not
really an issue, as it is a very easy migration. About how to migrate,
one can just do the same as done in `SerialPort.writeBytes1()` (notice
the '1' in the name). Just check if returned value matches the length of
the passed-in array. There is also a YAGNI approach (NOT RECOMMENDED!):
One could just use `writeBytes1()` instead. Which provides the old
behavior.

A more severe issue can arise because it potentially also breaks the ABI.
make sure to NOT mix usage of parts of jssc built before or after
this change! in most cases this shouldn't be an issue, as jssc unpacks
the correct version from its JAR anyway. BUT FOR THOSE WHO DID SEPARATE
THE JAVA CODE FROM THE NATIVE CODE AND INSTALL IT SEPARATELY, THEY
POTENTIALLY CAN END UP IN TROUBLE WHEN NOT TAKING CARE HERE.

In addition, this commit also contains some minor cleanups and fixes
some compiler warnings.

Related:
- https://github.com/java-native/jssc/issues/129

(wip @ 17cc247b69cdae978ad7192a21548162fe118f2a)